+++ lenstools/simulations/io.py
@@ -127,7 +127,7 @@ def saveFITS(self,filename,double_precision):
 		if double_precision:
 			hdu = fits.PrimaryHDU(self.data)
 		else:
-			hdu = fits.PrimaryHDU(self.data.astype(np.float32))
+			hdu = fits.PrimaryHDU(self.data.astype(np.float64))
 			
 	elif self.space=="fourier":
